[HumanBenchmark][Backend][Events] Refactor Events
